+++ ChangeLog.md
@@ -4,7 +4,34 @@ This file is a running history of fixes and improvements from version 7.0
 onwards. Changes are documented for the newest version first. Within each
 version, classes are listed alphabetically.
 
+# Version 8.0.533 Released 05/02/2021 03:05:40
+
+Build against the most recent WizardWrx .NET API library constellation.
+
+Going forward, this library will be periodically refreshed by building it
+against newer versions of the WizardWrx .NET API. If AssemblyInfo.cs and
+ChangeLog.md (this file) are unchanged, it is safe to assume that the only
+change is a refresh of its dependencies.
+
+# Version 8.0, Released 03/05/2021 18:55:45
+
+This update of the `WizardWrx.ConsoleAppAids3` library defines a new method,
+`LoadBasicErrorMessages`, that loads the appropriate string described by the
+following table.
+
+|Message Text                      |Common Resource| Common Code |Code Value|
+|----------------------------------|---------------|-------------|----------|
+|The task completed successfully.  |ERRMSG_SUCCESS |ERROR_SUCCESS|0x00000000|
+|A run-time exception has occurred.|ERRMSG_RUNTIME |ERROR_RUNTIME|0x00000001|
+
+Additional error messages can be loaded by passing in an optional array of
+strings, which must follow the order of exit code assignments, which do double
+duty as an index into the arry of messages. If you need only those two, both of
+which are required to enable other parts of the library to work correctly, pass
+in a NULL.
+
 # Version 7.2
+
 This update of the `WizardWrx.ConsoleAppAids3` library resolves an issue that
 allowed a calling routine to instantiate an uninitialized, utterly useless
 instance of the `ConsoleAppStateManager` object. In so doing, the bug breaks a